Make RouteState Members private
Organize files by State and Actions
