+++ source/BlazorState/Features/Routing/Actions/ChangeRoute/ChangeRouteRequest.cs
